refactor(frontend): distributes the page into multiple components and improves the style
Before, the <Table/> was a single component. Now a state manager (context API) has been added and
the table has been rewritten into 4 smaller components. There have also been improvements to the
file feedback mechanics. Added a footer. Added 'loading' for when the API request is being made, and
a toast system for when we get a request response. Styling is also improved overall, but still with
pure CSS.